Allow chart data inside shortcode with YAML format
This commit enables the following approach

[chartjs]
type: bar
data:
    labels: [Jan, Feb, Mar]
    datasets:
      - label: Bars Other Title
        data: [1,2,4]
[/chartjs]

Significant limitation of this approach is that it (currently) requires disabling all Markdown
processing on the page to work. Would need some raw page processing to fix this, see
the TODO in the source code. It does at least try to detect and warn you off if you try
this approach without disabling md processing.

Huge benefit is that the strings used map 1-1 to the Chart.js library, so any Chart.js
parameters/datasets/options can be utilized. Also, folks can directly reference the
Chart.js documentation, no need to read the readme for this plugin to understand how
the options here map to chart.js options

In order to convert YAML to JSON (faithfully) you need a proper YAML parser. Rather than
asking folks to install a PECL extension, I went with the battle-tested Symphony YAML
parser. The license is MIT, so it should not conflict with this plugin’s license. I chose to
follow the format used in multiple other plugins (and kind of documented at
https://github.com/getgrav/grav/issues/32 ) of embedding the composer “vendor” folder
directly into the git repo. Not what you typically do, but apparently that is the grav way.
I thought I would find a native YAML parser inside Grav (and it’s probably there somewhere)
but I gave up after a bit of looking and pulled in one I trusted.

My thoughts on using YAML (instead of JSON) inside the shortcode - A shortcode
is intended to allow for both quick writing and quick reading. IMO using raw JSON as the
contents would have worsened both the quick-writing and quick-reading nature of this
shortcode extension. As YAML is a pure superset of JSON, any JSON data can be
faithfully reproduced using YAML. If a user has any desire at all to go to raw JSON,
they can always paste that JSON directly into the markdown and bypass the shortcode
mechanism entirely. IMO the only negative of the YAML is that it makes it harder to
copy/paste chart examples into grav using this plugin, but I hope to address that by
adding some other functionality later.